+++ apps/site/assets/ts/helpers/__tests__/use-line-suspensions-test.ts
@@ -1,4 +1,4 @@
-import currentLineSuspensions from "./../use-line-suspensions";
+import currentLineSuspensions from "../line-suspensions";
 
 // TODO: Add tests
 test("currentLineSuspensions", () => {
